+++ lading/src/generator/file_gen/traditional.rs
@@ -78,7 +78,7 @@ fn default_rotation() -> bool {
     true
 }
 
-#[derive(Debug, Deserialize, Serialize, PartialEq)]
+#[derive(Debug, Deserialize, Serialize, PartialEq, Clone)]
 #[serde(deny_unknown_fields)]
 /// Configuration of [`FileGen`]
 pub struct Config {