+++ lading/src/generator/unix_datagram.rs
@@ -42,7 +42,7 @@ fn maximum_block_size() -> Byte {
     Byte::from_u64_with_unit(8_192, Unit::B).expect("catastrophic programming bug")
 }
 
-#[derive(Debug, Deserialize, Serialize, PartialEq)]
+#[derive(Debug, Deserialize, Serialize, PartialEq, Clone)]
 #[serde(deny_unknown_fields)]
 /// Configuration of this generator.
 pub struct Config {