+++ lading/src/target_metrics.rs
@@ -19,7 +19,7 @@ pub enum Error {
     Prometheus(prometheus::Error),
 }
 
-#[derive(Debug, Deserialize, PartialEq, Eq)]
+#[derive(Debug, Deserialize, PartialEq, Eq, Clone)]
 #[serde(rename_all = "snake_case")]
 #[serde(deny_unknown_fields)]
 /// Configuration for [`Server`]