+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+## Added
+- Lading now supports configuration directories. If the config path is
+  discovered to be a directory lading will merge all .yaml files in that
+  directory. Configuration files must be disjoint, else parsing will fail and
+  lading will not start up.
+
 ## [0.29.2]
 ## Added
 - Introduced multi-format writing for capture files. When format is 'multi' data